vocab_project/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.0.0
whitenoise>=6.6.0
Pillow>=10.0.0
